menambahkan dokumentasi

+++ database/migrations/2025_04_10_162670_create_pegawais_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('pegawais', function (Blueprint $table) {
             $table->id();
-            $table->integer('nip');
+            $table->bigInteger('nip');
             $table->string('nama');
             $table->string('tempat_lahir');
             $table->text('alamat');
